Add ability to set number of threads
Allow setting the number of threads to compress with using cd_values, and decompression threads using the HDF_FILTER_THREADS environment variable (since number of threads is not stored in the metadata and thus not available to the decompressor). Also move compressor setup to blosc_set_local() so it's only done once.